backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
email-validator==2.1.1
slowapi==0.1.8
